dbProject to TS
